- Fixed bug where IGNs were only world-specific  - Corrected min password size  - Fixed bug where experience gains would error 38 the client  - Overhauled the database schema; made everything consistent  - Removed the storage_items table, I feel that having basically the same schema in two places was a bad idea  - Fixed pet IDs to be 64-bit instead of 32-bit  - Zeroed out the SQL files because I can
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/branches/buis_vana_overhaul@3086 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/PetHandler.cpp
@@ -30,7 +30,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #include "SkillConstants.h"
 
 void PetHandler::handleMovement(Player *player, PacketReader &packet) {
-	int32_t petid = (int32_t)packet.get<int64_t>();
+	int64_t petid = packet.get<int64_t>();
 	Pet *pet = player->getPets()->getPet(petid);
 	if (pet == nullptr) {
 		// Hacking
@@ -43,7 +43,7 @@ void PetHandler::handleMovement(Player *player, PacketReader &packet) {
 }
 
 void PetHandler::handleChat(Player *player, PacketReader &packet) {
-	int32_t petid = (int32_t)packet.get<int64_t>();
+	int64_t petid = packet.get<int64_t>();
 	if (player->getPets()->getPet(petid) == nullptr) {
 		// Hacking
 		return;
@@ -153,7 +153,7 @@ void PetHandler::handleFeed(Player *player, PacketReader &packet) {
 }
 
 void PetHandler::handleCommand(Player *player, PacketReader &packet) {
-	int32_t petid = (int32_t)packet.get<int64_t>();
+	int64_t petid = packet.get<int64_t>();
 	Pet *pet = player->getPets()->getPet(petid);
 	if (pet == nullptr) {
 		// Hacking
@@ -174,7 +174,7 @@ void PetHandler::handleCommand(Player *player, PacketReader &packet) {
 }
 
 void PetHandler::handleConsumePotion(Player *player, PacketReader &packet) {
-	int32_t petid = (int32_t)packet.get<int64_t>();
+	int64_t petid = packet.get<int64_t>();
 	Pet *pet = player->getPets()->getPet(petid);
 	if (pet == nullptr || !pet->isSummoned() || player->getStats()->getHp() == 0) {
 		// Hacking
